routes/scan/scan_service.py
import os
from dotenv import load_dotenv
from fastapi import UploadFile, File

load_dotenv()

'''
 This is the service layer for the scan controller.
 It contains the business logic for virus and malware scanning and uploading files.
'''

def scan_file(file: UploadFile = File(...)):
    """
    Scan a file for viruses and malware.

    1. Check if the file is a valid file type.
    2. Check if the file is under 32MB.
    3. Scan the file for viruses and malware.
    4. If the file is clean, return the filename.
    """

    # Check if the file size exceeds 32MB
    if file.size > 32 * 1024 * 1024:  # 32MB
        return {"error": "File size exceeds 32MB limit."}

    # Here you would implement your virus and malware scanning logic
    # For now, we will just return the filename as a placeholder
    return {"filename": file.filename}